Let the builder fetch built assets of dev branches
We want to keep build files outside of git as much as possible.
A method to achieve this would be to only include build files in tagged
release commits were they don't interfere with development.

However, as each branch is tested by requiring it with composer, there
needs to be another solution to make sure that the files for these
branches are built as well. This is achieved here by having the builder
fetch the branch and build the assets, pass them to the build context
in a directory and then during build move all items in that directory to
the respective plugin installation directories.

To be able to do that the composer install command needed to be split up
into installation without running scripts and only running the scripts,
so that the assets can be moved to the right location in between.